requirements.txt
# langchain==0.0.284
# python-dotenv==1.0.0
# streamlit==1.22.0
# unstructured==0.9.2
# tiktoken==0.7.0
# faiss-cpu==1.8.0
# libmagic==1.0
# python-magic==0.4.27
# python-magic-bin==0.4.14
# OpenAI==0.28.0


# UI
streamlit==1.52.2

# Environment variables
python-dotenv==1.2.1

# LangChain core ecosystem (modern, stable)
langchain==1.2.3
langchain-core==1.2.7
langchain-community==0.4.1
langchain-groq==1.1.1
langchain-text-splitters==1.1.0

# Vector database
chromadb==1.4.0

# Embeddings (remote, no torch)
# huggingface-hub==0.20.3
sentence-transformers
torch

# Web loading
beautifulsoup4==4.14.3
requests==2.32.5

# Streamlit protobuf compatibility (CRITICAL)
protobuf==3.20.3
